+++ test/protractor/spec/lforms_formdata.spec.js
@@ -158,7 +158,6 @@ describe('formdata: ', function() {
         expect(formData.code).toBe("54127-6N");
         expect(formData.name).toBe("USSG-FHT, (with mock-up items for skip logic demo)");
         expect(formData.template).toBe("table");
-        expect(formData.type).toBe("LOINC");
         expect(formData.templateOptions.showFormHeader).toBe(true);
         expect(formData.templateOptions.showQuestionCode).toBe(false);
         expect(formData.items[0].question).toBe("Your health information");
@@ -199,7 +198,6 @@ describe('formdata: ', function() {
         expect(formData.code).toBe("54127-6N");
         expect(formData.name).toBe("USSG-FHT, (with mock-up items for skip logic demo)");
         expect(formData.template).toBe("table");
-        expect(formData.type).toBe("LOINC");
         expect(formData.templateOptions.showFormHeader).toBe(true);
         expect(formData.templateOptions.showQuestionCode).toBe(false);
         expect(formData.items[0].question).toBe("Your health information");